Database update: keeping size_t vector instead of ints

+++ IDatabase/Database.cpp
@@ -6,13 +6,13 @@ Database::Database(sql *connect):connection(connect) {}
 
 Database::~Database()  {}
 
-bool Database::Insert(int id, const std::vector<int> &res)  {
+bool Database::Insert(int id, const std::vector<size_t> &res)  {
     return true;
 }
 bool Database::Insert(Result a)  {
     return true;
 }
 
-std::vector<int> Select(int id) {
-    return std::vector<int> ();
+std::vector<size_t> Select(int id) {
+    return std::vector<size_t> ();
 }
