+++ IDatabase/TestDatabase/TestDatabase.cpp
@@ -2,7 +2,7 @@
 
 //int TestDatabase::TestDatabase() {}
 
-bool TestDatabase::Insert(int u_id, const std::vector<int>& t_vec) {
+bool TestDatabase::Insert(int u_id, const std::vector<size_t>& t_vec) {
 
     if ( this->id == u_id ) return false;
 
@@ -23,8 +23,8 @@ bool TestDatabase::Insert( Result a) {
 
 TestDatabase::~TestDatabase() {}
 
-std::vector<int> TestDatabase::Select(int u_id) {
-    if ( this->id != u_id ) return std::vector<int>();
+std::vector<size_t> TestDatabase::Select(int u_id) {
+    if ( this->id != u_id ) return std::vector<size_t>();
 
     return this->data;
 }
